Add C11 support module
